% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_birds_doc.R
\docType{data}
\name{cap_birds}
\alias{cap_birds}
\title{Bird Community Observations (2000 - 2019), Central Arizona - Phoenix LTER}
\format{
A tibble with 172,175 rows and 14 variables
\describe{
\item{survey_id}{a number denoting the unique identity of the bird survey}
\item{site_code}{a character denoting the identity of the survey location}
\item{survey_date}{a date denoting the date of observation}
\item{time_start}{a time denoting the start time of the observation period}
\item{time_end}{a time denoting the end time of the observation period}
\item{common_name}{a character denoting the common name of the species observed}
\item{distance}{a character denoting the distance of the observed birds from the observer in meters}
\item{bird_count}{a number denoting the count of birds observed during the survey period}
\item{seen}{a boolean denoting whether the bird was seen by the observer}
\item{heard}{a boolean denoting whether the bird was heard by the observer}
\item{direction}{a character denoting the direction of the bird from the observation point}
\item{location_type}{a character denoting the physical or research context of the site}
\item{lat}{a number denoting the latitude of the survey site in degrees}
\item{long}{a number denoting the longitude of the survey site in degrees}
}
}
\source{
{Warren, P., S.B. Lerman, H. Bateman, M. Katti, and E. Shochat. 2022. Point-count bird censusing: long-term monitoring of bird abundance and diversity in central Arizona-Phoenix, ongoing since 2000 ver 19. Environmental Data Initiative.}
\url{https://doi.org/10.6073/pasta/4fca7c8a6cd56a6abed9834aca72e764}
}
\usage{
cap_birds
}
\description{
Bird count data collected in different seasons using a standardized point-count protocol.  In 2000, sites were visited twice. In 2001 each site was visited in each season by three independent observers. In 2005, sites were visited in three seasons, and from 2006 - 2019, only in spring and winter.
}
\keyword{datasets}
