% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lterdatasampler-package.R
\docType{package}
\name{lterdatasampler-package}
\alias{lterdatasampler}
\alias{lterdatasampler-package}
\title{lterdatasampler: Educational Dataset Examples from the Long Term Ecological Research Program}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Curated datasets from US Long Term Ecological Research sites.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lter/lterdatasampler}
  \item Report bugs at \url{https://github.com/lter/lterdatasampler/issues}
}

}
\author{
\strong{Maintainer}: Allison Horst \email{ahorst@ucsb.edu} (\href{https://orcid.org/0000-0002-6047-5564}{ORCID})

Authors:
\itemize{
  \item Julien Brun \email{brun@nceas.ucsb.edu} (\href{https://orcid.org/0000-0002-7751-6238}{ORCID})
}

Other contributors:
\itemize{
  \item Sam Guo (https://github.com/TokyoExpress) [contributor]
  \item Adhitya Logan (https://github.com/adhil0) [contributor]
  \item Lian Ran (https://github.com/liaaaaran) [contributor]
  \item Sophia Sternberg (https://github.com/sophiasternberg) [contributor]
  \item Karen Zhao (https://github.com/karenezhao) [contributor]
}

}
\keyword{internal}
