% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luq_streamchem_doc.R
\docType{data}
\name{luq_streamchem}
\alias{luq_streamchem}
\title{LUQ Stream Chemistry Data for Quebrada Sonadora (QS) site}
\format{
A tibble with 4,092 rows and 4 variables:
\describe{
\item{common_name}{a character denoting the species common name)}
\item{site}{a character denoting research site}
\item{sex}{a character denoting lizard sex; M = Male, F = Female, J = Juvenile}
\item{tail}{a character denoting lizard tail condition; B = broken, W = whole}
}
}
\source{
{McDowell, W. 2017. Chemistry of stream water from the Luquillo Mountains ver 4923051. Environmental Data Initiative.} \url{https://doi.org/10.6073/pasta/f9df56348f510da0113b1e6012fa2967}
}
\usage{
luq_streamchem
}
\description{
Need to drop some columns
}
\keyword{datasets}
