% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lterdatasampler-package.R
\docType{package}
\name{lterdatasampler-package}
\alias{lterdatasampler}
\alias{lterdatasampler-package}
\title{lterdatasampler: Educational Dataset Examples from the Long Term Ecological Research Program}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Curated datasets from US Long Term Ecological Research sites. References: Gregory, S.V. and I. Arismendi (2020)\doi{10.6073/pasta/7c78d662e847cdbe33584add8f809165}, Shaver, G. (2019)\doi{10.6073/pasta/ce0f300cdf87ec002909012abefd9c5c}, Juice, S. and T. Fahey (2019)\doi{10.6073/pasta/0ade53ede9a916a36962799b2407097e}, EBlair, J (2021)\doi{10.6073/pasta/9c641b35695abc5889edd64c3950517f}, McDowell, W. (2021)\doi{10.6073/pasta/0a09f5aa2e6f11451553c92b102279a6}, Anderson, L. and D. Robertson (2020)\doi{10.6073/pasta/e3ff85971d817e9898bb8a83fb4c3a8b}, Magnuson, J.J., S.R. Carpenter, and E.H. Stanley (2021)\doi{10.6073/pasta/ab31f2489ee436beb73fc8f1d0213d97}, Whipple, A. and Niwot Ridge LTER (2020)\doi{10.6073/pasta/9f95baf55f98732f47a8844821ff690d}, Johnson, D. (2019)\doi{10.6073/pasta/4c27d2e778d3325d3830a5142e3839bb}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lter/lterdatasampler}
  \item Report bugs at \url{https://github.com/lter/lterdatasampler/issues}
}

}
\author{
\strong{Maintainer}: Allison Horst \email{ahorst@ucsb.edu} (\href{https://orcid.org/0000-0002-6047-5564}{ORCID})

Authors:
\itemize{
  \item Julien Brun \email{brun@nceas.ucsb.edu} (\href{https://orcid.org/0000-0002-7751-6238}{ORCID})
}

Other contributors:
\itemize{
  \item Sam Guo (https://github.com/TokyoExpress) [contributor]
  \item Adhitya Logan (https://github.com/adhil0) [contributor]
  \item Lian Ran (https://github.com/liaaaaran) [contributor]
  \item Sophia Sternberg (https://github.com/sophiasternberg) [contributor]
  \item Karen Zhao (https://github.com/karenezhao) [contributor]
}

}
\keyword{internal}
